fix(suite): Fix sidebar re-rendering
